Remove sub-directory source and move files to root directory Fix package includes
