+++ .envrc
@@ -1,0 +1,1 @@
+export GOPATH=`(cd ../../;pwd)`
