--- go/src/prclient/prclient.go
@@ -1,561 +1,0 @@
-// PavedRoad APIs follow kubernetes conventions
-// /api/v1/namespace/{namespace}/resourcetype default
-// The default namespace is pavedroad.io
-//
-// Verbs
-// ---------
-// GET:      Return a resource
-// POST:     Create a new token
-// PUT:      Update a specific toke
-// PATCH:    Partial update of a specific toke
-// DELETE:   Delete a specific toke
-//
-// Resources
-// ---------
-// prToken:  A token for accessing 3rd party services
-//
-// UUID are used to identify all resources
-//
-
-package prclient
-
-import (
-	"bytes"
-	"context"
-	"encoding/json"
-	"fmt"
-	"github.com/google/go-querystring/query"
-	"io"
-	"io/ioutil"
-	"net/http"
-	"net/url"
-	"reflect"
-	"strconv"
-	"strings"
-	"sync"
-)
-
-const (
-	headerOTP = "X-PavedRoad-OTP"
-
-	mediaTypeV3      = "application/vnd.pavedroad.v3+json"
-	defaultMediaType = "application/octet-stream"
-
-	apiVersion         string = "/api/v1/"
-	namespaceID        string = "namespace/"
-	defaultNamespace   string = "pavedroad.io/"
-	tokenResource      string = "prTokens"
-	tokenResourceList  string = "prTokensLIST"
-	gitHubResource     string = "prGitHub"
-	userResource       string = "prUser"
-	repositoryResource string = "prRepository"
-	mapperResource      string = "prUserIdMappers"
-	mapperResourceList  string = "prUserIdMappersLIST"
-	uid                string = "{uid}"
-	cred               string = "{cred}"
-
-	// TODO: Turn this into a function
-	defaultBaseURL = "https://api.pavedroad.io" + apiVersion + namespaceID + defaultNamespace
-	uploadBaseURL  = "https://uploads.pavedroad.io/"
-	userAgent      = "prclient"
-)
-
-// A Client manages communication with the PavedRoad API.
-type Client struct {
-	clientMu sync.Mutex   // clientMu protects the client during calls that modify the CheckRedirect func.
-	client   *http.Client // HTTP client used to communicate with the API.
-
-	// Base URL for API requests. Defaults to the public PavedRoad API, but can be
-	// set to a domain endpoint to use with PavedRoad Enterprise. BaseURL should
-	// always be specified with a trailing slash.
-	BaseURL *url.URL
-
-	// Base URL for uploading files.
-	UploadURL *url.URL
-
-	// User agent used when communicating with the PavedRoad API.
-	UserAgent string
-
-	common service // Reuse a single struct instead of allocating one for each service on the heap.
-
-	// Services used for talking to different parts of the PavedRoad API.
-	Token *TokensService
-	UserIdMapper *UserIdMappersService
-}
-
-type service struct {
-	client *Client
-}
-
-// ListOptions specifies the optional parameters to various List methods that
-// support pagination.
-type ListOptions struct {
-	// For paginated result sets, page of results to retrieve.
-	Page int `url:"page,omitempty"`
-
-	// For paginated result sets, the number of results to include per page.
-	PerPage int `url:"per_page,omitempty"`
-}
-
-// UploadOptions specifies the parameters to methods that support uploads.
-type UploadOptions struct {
-	Name      string `url:"name,omitempty"`
-	Label     string `url:"label,omitempty"`
-	MediaType string `url:"-"`
-}
-
-// RawType represents type of raw format of a request instead of JSON.
-type RawType uint8
-
-const (
-	// Diff format.
-	Diff RawType = 1 + iota
-	// Patch format.
-	Patch
-)
-
-// RawOptions specifies parameters when user wants to get raw format of
-// a response instead of JSON.
-type RawOptions struct {
-	Type RawType
-}
-
-// addOptions adds the parameters in opt as URL query parameters to s. opt
-// must be a struct whose fields may contain "url" tags.
-func addOptions(s string, opt interface{}) (string, error) {
-	v := reflect.ValueOf(opt)
-	if v.Kind() == reflect.Ptr && v.IsNil() {
-		return s, nil
-	}
-
-	u, err := url.Parse(s)
-	if err != nil {
-		return s, err
-	}
-
-	qs, err := query.Values(opt)
-	if err != nil {
-		return s, err
-	}
-
-	u.RawQuery = qs.Encode()
-	return u.String(), nil
-}
-
-// NewClient returns a new PavedRoad API client. If a nil httpClient is
-// provided, a new http.Client will be used. To use API methods which require
-// authentication, provide an http.Client that will perform the authentication
-// for you (such as that provided by the golang.org/x/oauth2 library).
-func NewClient(httpClient *http.Client) *Client {
-	if httpClient == nil {
-		httpClient = &http.Client{}
-	}
-	baseURL, _ := url.Parse(defaultBaseURL)
-	uploadURL, _ := url.Parse(uploadBaseURL)
-
-	c := &Client{client: httpClient, BaseURL: baseURL, UserAgent: userAgent, UploadURL: uploadURL}
-	c.common.client = c
-	c.Token = (*TokensService)(&c.common)
-	c.UserIdMapper = (*UserIdMappersService)(&c.common)
-	//c.Users = (*UsersService)(&c.common)
-	return c
-}
-
-// NewRequest creates an API request. A relative URL can be provided in urlStr,
-// in which case it is resolved relative to the BaseURL of the Client.
-// Relative URLs should always be specified without a preceding slash. If
-// specified, the value pointed to by body is JSON encoded and included as the
-// request body.
-func (c *Client) NewRequest(method, urlStr string, body interface{}) (*http.Request, error) {
-	if !strings.HasSuffix(c.BaseURL.Path, "/") {
-		return nil, fmt.Errorf("BaseURL must have a trailing slash, but %q does not", c.BaseURL)
-	}
-	u, err := c.BaseURL.Parse(urlStr)
-	if err != nil {
-		return nil, err
-	}
-
-	var buf io.ReadWriter
-	if body != nil {
-		buf = new(bytes.Buffer)
-		enc := json.NewEncoder(buf)
-		enc.SetEscapeHTML(false)
-		err := enc.Encode(body)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	req, err := http.NewRequest(method, u.String(), buf)
-	if err != nil {
-		return nil, err
-	}
-
-	if body != nil {
-		req.Header.Set("Content-Type", "application/json")
-	}
-	req.Header.Set("Accept", mediaTypeV3)
-	if c.UserAgent != "" {
-		req.Header.Set("User-Agent", c.UserAgent)
-	}
-	return req, nil
-}
-
-// NewUploadRequest creates an upload request. A relative URL can be provided in
-// urlStr, in which case it is resolved relative to the UploadURL of the Client.
-// Relative URLs should always be specified without a preceding slash.
-func (c *Client) NewUploadRequest(urlStr string, reader io.Reader, size int64, mediaType string) (*http.Request, error) {
-	if !strings.HasSuffix(c.UploadURL.Path, "/") {
-		return nil, fmt.Errorf("UploadURL must have a trailing slash, but %q does not", c.UploadURL)
-	}
-	u, err := c.UploadURL.Parse(urlStr)
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequest("POST", u.String(), reader)
-	if err != nil {
-		return nil, err
-	}
-	req.ContentLength = size
-
-	if mediaType == "" {
-		mediaType = defaultMediaType
-	}
-	req.Header.Set("Content-Type", mediaType)
-	req.Header.Set("Accept", mediaTypeV3)
-	req.Header.Set("User-Agent", c.UserAgent)
-	return req, nil
-}
-
-// Response is a PavedRoad API response. This wraps the standard http.Response
-// returned from PavedRoad and provides convenient access to things like
-// pagination links.
-type Response struct {
-	*http.Response
-
-	// These fields provide the page values for paginating through a set of
-	// results. Any or all of these may be set to the zero value for
-	// responses that are not part of a paginated set, or for which there
-	// are no additional pages.
-
-	NextPage  int
-	PrevPage  int
-	FirstPage int
-	LastPage  int
-}
-
-// newResponse creates a new Response for the provided http.Response.
-// r must not be nil.
-func newResponse(r *http.Response) *Response {
-	response := &Response{Response: r}
-	response.populatePageValues()
-	return response
-}
-
-// populatePageValues parses the HTTP Link response headers and populates the
-// various pagination link values in the Response.
-func (r *Response) populatePageValues() {
-	if links, ok := r.Response.Header["Link"]; ok && len(links) > 0 {
-		for _, link := range strings.Split(links[0], ",") {
-			segments := strings.Split(strings.TrimSpace(link), ";")
-
-			// link must at least have href and rel
-			if len(segments) < 2 {
-				continue
-			}
-
-			// ensure href is properly formatted
-			if !strings.HasPrefix(segments[0], "<") || !strings.HasSuffix(segments[0], ">") {
-				continue
-			}
-
-			// try to pull out page parameter
-			url, err := url.Parse(segments[0][1 : len(segments[0])-1])
-			if err != nil {
-				continue
-			}
-			page := url.Query().Get("page")
-			if page == "" {
-				continue
-			}
-
-			for _, segment := range segments[1:] {
-				switch strings.TrimSpace(segment) {
-				case `rel="next"`:
-					r.NextPage, _ = strconv.Atoi(page)
-				case `rel="prev"`:
-					r.PrevPage, _ = strconv.Atoi(page)
-				case `rel="first"`:
-					r.FirstPage, _ = strconv.Atoi(page)
-				case `rel="last"`:
-					r.LastPage, _ = strconv.Atoi(page)
-				}
-
-			}
-		}
-	}
-}
-
-// Do sends an API request and returns the API response. The API response is
-// JSON decoded and stored in the value pointed to by v, or returned as an
-// error if an API error has occurred. If v implements the io.Writer
-// interface, the raw response body will be written to v, without attempting to
-// first decode it.
-//
-// The provided ctx must be non-nil. If it is canceled or times out,
-// ctx.Err() will be returned.
-func (c *Client) Do(ctx context.Context, req *http.Request, v interface{}) (*Response, error) {
-	resp, err := c.client.Do(req)
-	if err != nil {
-		// If we got an error, and the context has been canceled,
-		// the context's error is probably more useful.
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		default:
-		}
-
-		// If the error type is *url.Error, sanitize its URL before returning.
-		if e, ok := err.(*url.Error); ok {
-			if url, err := url.Parse(e.URL); err == nil {
-				e.URL = sanitizeURL(url).String()
-				return nil, e
-			}
-		}
-
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	response := newResponse(resp)
-
-	err = CheckResponse(resp)
-	if err != nil {
-		// Special case for AcceptedErrors. If an AcceptedError
-		// has been encountered, the response's payload will be
-		// added to the AcceptedError and returned.
-		//
-		// Issue #1022
-		aerr, ok := err.(*AcceptedError)
-		if ok {
-			b, readErr := ioutil.ReadAll(resp.Body)
-			if readErr != nil {
-				return response, readErr
-			}
-
-			aerr.Raw = b
-			return response, aerr
-		}
-
-		return response, err
-	}
-
-	if v != nil {
-		if w, ok := v.(io.Writer); ok {
-			io.Copy(w, resp.Body)
-		} else {
-			decErr := json.NewDecoder(resp.Body).Decode(v)
-			if decErr == io.EOF {
-				decErr = nil // ignore EOF errors caused by empty response body
-			}
-			if decErr != nil {
-				err = decErr
-			}
-		}
-	}
-
-	return response, err
-}
-
-/*
-An ErrorResponse reports one or more errors caused by an API request.
-
-TODO:(tbd)PavedRoad API docs: https://developer.pavedroad.io/v3/#client-errors
-*/
-type ErrorResponse struct {
-	Response *http.Response // HTTP response that caused this error
-	Message  string         `json:"message"` // error message
-	Errors   []Error        `json:"errors"`  // more detail on individual errors
-	// Block is only populated on certain types of errors such as code 451.
-	Block *struct {
-		Reason    string     `json:"reason,omitempty"`
-		CreatedAt *Timestamp `json:"created_at,omitempty"`
-	} `json:"block,omitempty"`
-	// Most errors will also include a documentation_url field pointing
-	// to some content that might help you resolve the error, see
-	// TODO:(tbd)  https://developer.pavedroad.io/v3/#client-errors
-	DocumentationURL string `json:"documentation_url,omitempty"`
-}
-
-func (r *ErrorResponse) Error() string {
-	return fmt.Sprintf("%v %v: %d %v %+v",
-		r.Response.Request.Method, sanitizeURL(r.Response.Request.URL),
-		r.Response.StatusCode, r.Message, r.Errors)
-}
-
-// AcceptedError occurs when PavedRoad returns 202 Accepted response with an
-// empty body, which means a job was scheduled on the PavedRoad side to process
-// the information needed and cache it.
-// Technically, 202 Accepted is not a real error, it's just used to
-// indicate that results are not ready yet, but should be available soon.
-// The request can be repeated after some time.
-type AcceptedError struct {
-	// Raw contains the response body.
-	Raw []byte
-}
-
-func (*AcceptedError) Error() string {
-	return "job scheduled on PavedRoad side; try again later"
-}
-
-// sanitizeURL redacts the client_secret parameter from the URL which may be
-// exposed to the user.
-func sanitizeURL(uri *url.URL) *url.URL {
-	if uri == nil {
-		return nil
-	}
-	params := uri.Query()
-	if len(params.Get("client_secret")) > 0 {
-		params.Set("client_secret", "REDACTED")
-		uri.RawQuery = params.Encode()
-	}
-	return uri
-}
-
-/*
-An Error reports more details on an individual error in an ErrorResponse.
-These are the possible validation error codes:
-
-    missing:
-        resource does not exist
-    missing_field:
-        a required field on a resource has not been set
-    invalid:
-        the formatting of a field is invalid
-    already_exists:
-        another resource has the same valid as this field
-    custom:
-        some resources return this, additional information is
-        set in the Message field of the Error
-
-PavedRoad API docs: https://developer.pavedroad.io/v1/#client-errors
-*/
-type Error struct {
-	Resource string `json:"resource"` // resource on which the error occurred
-	Field    string `json:"field"`    // field on which the error occurred
-	Code     string `json:"code"`     // validation error code
-	Message  string `json:"message"`  // Message describing the error. Errors with Code == "custom" will always have this set.
-}
-
-func (e *Error) Error() string {
-	return fmt.Sprintf("%v error caused by %v field on %v resource",
-		e.Code, e.Field, e.Resource)
-}
-
-// CheckResponse checks the API response for errors, and returns them if
-// present. A response is considered an error if it has a status code outside
-// the 200 range or equal to 202 Accepted.
-// API error responses are expected to have either no response
-// body, or a JSON response body that maps to ErrorResponse. Any other
-// response body will be silently ignored.
-//
-// *AcceptedError for 202 Accepted status codes,
-// and *TwoFactorAuthError for two-factor authentication errors.
-func CheckResponse(r *http.Response) error {
-	if r.StatusCode == http.StatusAccepted {
-		return &AcceptedError{}
-	}
-	if c := r.StatusCode; 200 <= c && c <= 299 {
-		return nil
-	}
-	errorResponse := &ErrorResponse{Response: r}
-	data, err := ioutil.ReadAll(r.Body)
-	if err == nil && data != nil {
-		json.Unmarshal(data, errorResponse)
-	}
-	return errorResponse
-}
-
-// parseBoolResponse determines the boolean result from a PavedRoad API response.
-// Several PavedRoad API methods return boolean responses indicated by the HTTP
-// status code in the response (true indicated by a 204, false indicated by a
-// 404). This helper function will determine that result and hide the 404
-// error if present. Any other error will be returned through as-is.
-func parseBoolResponse(err error) (bool, error) {
-	if err == nil {
-		return true, nil
-	}
-
-	if err, ok := err.(*ErrorResponse); ok && err.Response.StatusCode == http.StatusNotFound {
-		// Simply false. In this one case, we do not pass the error through.
-		return false, nil
-	}
-
-	// some other real error occurred
-	return false, err
-}
-
-// RoundTrip implements the RoundTripper interface.
-func (t *BasicAuthTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	// To set extra headers, we must make a copy of the Request so
-	// that we don't modify the Request we were given. This is required by the
-	// specification of http.RoundTripper.
-	//
-	// Since we are going to modify only req.Header here, we only need a deep copy
-	// of req.Header.
-	req2 := new(http.Request)
-	*req2 = *req
-	req2.Header = make(http.Header, len(req.Header))
-	for k, s := range req.Header {
-		req2.Header[k] = append([]string(nil), s...)
-	}
-
-	req2.SetBasicAuth(t.Username, t.Password)
-	if t.OTP != "" {
-		req2.Header.Set(headerOTP, t.OTP)
-	}
-	return t.transport().RoundTrip(req2)
-}
-
-// BasicAuthTransport is an http.RoundTripper that authenticates all requests
-// using HTTP Basic Authentication with the provided username and password. It
-// additionally supports users who have two-factor authentication enabled.
-type BasicAuthTransport struct {
-	Username string // PavedRoad username
-	Password string // PavedRoad password
-	OTP      string // one-time password for users with two-factor auth enabled
-
-	// Transport is the underlying HTTP transport to use when making requests.
-	// It will default to http.DefaultTransport if nil.
-	Transport http.RoundTripper
-}
-
-// Client returns an *http.Client that makes requests that are authenticated
-// using HTTP Basic Authentication.
-func (t *BasicAuthTransport) Client() *http.Client {
-	return &http.Client{Transport: t}
-}
-
-func (t *BasicAuthTransport) transport() http.RoundTripper {
-	if t.Transport != nil {
-		return t.Transport
-	}
-	return http.DefaultTransport
-}
-
-// Bool is a helper routine that allocates a new bool value
-// to store v and returns a pointer to it.
-func Bool(v bool) *bool { return &v }
-
-// Int is a helper routine that allocates a new int value
-// to store v and returns a pointer to it.
-func Int(v int) *int { return &v }
-
-// Int64 is a helper routine that allocates a new int64 value
-// to store v and returns a pointer to it.
-func Int64(v int64) *int64 { return &v }
-
-// String is a helper routine that allocates a new string value
-// to store v and returns a pointer to it.
-func String(v string) *string { return &v }
