--- go/src/prclient/timestamp.go
@@ -1,34 +1,0 @@
-package prclient
-
-import (
-	"strconv"
-	"time"
-)
-
-// Timestamp represents a time that can be unmarshalled from a JSON string
-// formatted as either an RFC3339 or Unix timestamp.
-type Timestamp struct {
-	time.Time
-}
-
-func (t Timestamp) String() string {
-	return t.Time.String()
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface.
-// Time is expected in RFC3339 or Unix format.
-func (t *Timestamp) UnmarshalJSON(data []byte) (err error) {
-	str := string(data)
-	i, err := strconv.ParseInt(str, 10, 64)
-	if err == nil {
-		t.Time = time.Unix(i, 0)
-	} else {
-		t.Time, err = time.Parse(`"`+time.RFC3339+`"`, str)
-	}
-	return
-}
-
-// Equal reports whether t and u are equal based on time.Equal
-func (t Timestamp) Equal(u Timestamp) bool {
-	return t.Time.Equal(u.Time)
-}
